+++ lib/src/watchers/listenable.dart
@@ -3,6 +3,8 @@ import 'package:flutter/foundation.dart';
 import '../framework.dart';
 
 class ListenableWatcher extends Watcher<Listenable> {
+  const ListenableWatcher();
+
   @override
   void init(Listenable observable, VoidCallback notify) {
     observable.addListener(notify);
@@ -11,5 +13,12 @@ class ListenableWatcher extends Watcher<Listenable> {
   @override
   void cancel(Listenable observable, VoidCallback notify) {
     observable.removeListener(notify);
+  }
+
+  @override
+  void dispose(Listenable observable) {
+    if (observable is ChangeNotifier) {
+      observable.dispose();
+    }
   }
 }